fix: unmarshal window field as float64 to handle JSON decimal from auth proxy
The auth proxy returns the window value as a JSON float (e.g. 35505099.0),
which Go's JSON decoder rejects when the target type is int64.

+++ pi5/vault.go
@@ -28,7 +28,7 @@ const (
 type SessionKeyResponse struct {
 	SessionKey string `json:"session_key"`
 	ExpiresAt  string `json:"expires_at"`
-	Window     int64  `json:"window"`
+	Window     float64 `json:"window"`
 }
 
 type AuditEntry struct {
